@@ -13,7 +13,10 @@
     "serve": "serve build"
   },
   "eslintConfig": {
-    "extends": "react-app"
+    "extends": [
+      "react-app",
+      "react-app/jest"
+    ]
   },
   "browserslist": {
     "production": [
@@ -31,25 +34,28 @@
     "bootstrap": "^4.4.1",
     "hex-rgb": "^5.0.0",
     "moment": "^2.24.0",
-    "react": "^17.0.1",
-    "react-dom": "^17.0.1",
+    "react": "^18.0.0",
+    "react-dom": "^18.0.0",
     "react-helmet": "^6.1.0",
     "react-moment": "^1.0.0",
     "reactstrap": "^8.4.1",
     "styled-components": "^5.2.1"
   },
   "devDependencies": {
-    "@testing-library/jest-dom": "^5.11.6",
-    "@testing-library/react": "^12.0.0",
-    "@testing-library/user-event": "^13.2.1",
-    "@types/react": "^17.0.0",
-    "@types/react-dom": "^17.0.0",
+    "@testing-library/jest-dom": "^5.16.3",
+    "@testing-library/react": "^12.1.4",
+    "@testing-library/user-event": "^13.5.0",
+    "@types/jest": "^27.4.1",
+    "@types/node": "^16.11.26",
+    "@types/react": "^17.0.43",
+    "@types/react-dom": "^17.0.14",
     "@types/react-helmet": "^6.1.0",
     "@types/styled-components": "^5.1.7",
     "gh-pages": "^3.1.0",
     "node-sass": "^6.0.1",
-    "react-scripts": "^4.0.1",
-    "typescript": "^4.1.2",
-    "serve": "^12.0.0"
+    "react-scripts": "5.0.0",
+    "serve": "^13.0.2",
+    "typescript": "^4.6.3",
+    "web-vitals": "^2.1.4"
   }
 }
